chore: Move engine builder tests from turborepo-lib to turborepo-engine
Move 3,196 lines of EngineBuilder integration tests from turborepo-lib
to turborepo-engine where the production code lives.

## Changes

- Add dev-dependencies to turborepo-engine for test infrastructure
- Move all 47 test functions to turborepo-engine/src/builder.rs
- Delete turborepo-lib/src/engine/builder.rs (was tests-only)
- Remove `mod builder;` from turborepo-lib/src/engine/mod.rs

## Why

The tests were testing EngineBuilder which now lives in turborepo-engine.
Tests should be colocated with the code they test. This reduces
turborepo-lib by ~3,200 lines as part of the ongoing modularization.
