+++ crates/turborepo-engine/Cargo.toml
@@ -28,5 +28,10 @@ turborepo-types = { workspace = true }
 which = { workspace = true }
 
 [dev-dependencies]
+insta = { workspace = true }
 pretty_assertions = { workspace = true }
+serde_json = { workspace = true }
+tempfile = { workspace = true }
 test-case = { workspace = true }
+turborepo-lockfiles = { path = "../turborepo-lockfiles" }
+turborepo-turbo-json = { path = "../turborepo-turbo-json" }
